- Separated core data portion from view controller
